migrate away from common json codec and use marshal methods.
