--- pkg/cloudevents/codec/doc.go
@@ -1,5 +1,0 @@
-/*
-Package codec holds the encoder/decoder implementation for structured encodings using `application/json` of the
-whole CloudEvent.
-*/
-package codec
