+++ pkg/cloudevents/transport/amqp/codec_v02.go
@@ -4,11 +4,12 @@ import (
 	"fmt"
 
 	"github.com/cloudevents/sdk-go/pkg/cloudevents"
-	"github.com/cloudevents/sdk-go/pkg/cloudevents/codec"
 	"github.com/cloudevents/sdk-go/pkg/cloudevents/transport"
 )
 
 type CodecV02 struct {
+	CodecStructured
+
 	Encoding Encoding
 }
 
@@ -29,29 +30,10 @@ func (v CodecV02) Decode(msg transport.Message) (*cloudevents.Event, error) {
 	// only structured is supported as of v0.2
 	switch v.inspectEncoding(msg) {
 	case StructuredV02:
-		return v.decodeStructured(msg)
+		return v.decodeStructured(cloudevents.CloudEventsVersionV02, msg)
 	default:
 		return nil, transport.NewErrMessageEncodingUnknown("v02", TransportName)
 	}
-}
-
-func (v CodecV02) encodeStructured(e cloudevents.Event) (transport.Message, error) {
-	body, err := codec.JsonEncodeV02(e)
-	if err != nil {
-		return nil, err
-	}
-	return &Message{
-		Body:        body,
-		ContentType: cloudevents.ApplicationCloudEventsJSON,
-	}, nil
-}
-
-func (v CodecV02) decodeStructured(msg transport.Message) (*cloudevents.Event, error) {
-	m, ok := msg.(*Message)
-	if !ok {
-		return nil, fmt.Errorf("failed to convert transport.Message to amqp.Message")
-	}
-	return codec.JsonDecodeV02(m.Body)
 }
 
 func (v CodecV02) inspectEncoding(msg transport.Message) Encoding {
