+++ pkg/cloudevents/transport/amqp/codec_v02_test.go
@@ -29,11 +29,11 @@ func TestCodecV02_Encode(t *testing.T) {
 		"simple v2 default": {
 			codec: amqp.CodecV02{},
 			event: cloudevents.Event{
-				Context: &cloudevents.EventContextV02{
+				Context: cloudevents.EventContextV02{
 					Type:   "com.example.test",
 					Source: *source,
 					ID:     "ABC-123",
-				},
+				}.AsV02(),
 			},
 			want: &amqp.Message{
 				ContentType: cloudevents.ApplicationCloudEventsJSON,
@@ -52,7 +52,7 @@ func TestCodecV02_Encode(t *testing.T) {
 		"full v2 default": {
 			codec: amqp.CodecV02{},
 			event: cloudevents.Event{
-				Context: &cloudevents.EventContextV02{
+				Context: cloudevents.EventContextV02{
 					ID:          "ABC-123",
 					Time:        &now,
 					Type:        "com.example.test",
@@ -62,7 +62,7 @@ func TestCodecV02_Encode(t *testing.T) {
 					Extensions: map[string]interface{}{
 						"test": "extended",
 					},
-				},
+				}.AsV02(),
 				Data: map[string]interface{}{
 					"hello": "world",
 				},
@@ -90,11 +90,11 @@ func TestCodecV02_Encode(t *testing.T) {
 		"simple v2 structured": {
 			codec: amqp.CodecV02{Encoding: amqp.StructuredV02},
 			event: cloudevents.Event{
-				Context: &cloudevents.EventContextV02{
+				Context: cloudevents.EventContextV02{
 					Type:   "com.example.test",
 					Source: *source,
 					ID:     "ABC-123",
-				},
+				}.AsV02(),
 			},
 			want: &amqp.Message{
 				ContentType: cloudevents.ApplicationCloudEventsJSON,
@@ -113,7 +113,7 @@ func TestCodecV02_Encode(t *testing.T) {
 		"full v2 structured": {
 			codec: amqp.CodecV02{Encoding: amqp.StructuredV02},
 			event: cloudevents.Event{
-				Context: &cloudevents.EventContextV02{
+				Context: cloudevents.EventContextV02{
 					ID:          "ABC-123",
 					Time:        &now,
 					Type:        "com.example.test",
@@ -123,7 +123,7 @@ func TestCodecV02_Encode(t *testing.T) {
 					Extensions: map[string]interface{}{
 						"test": "extended",
 					},
-				},
+				}.AsV02(),
 				Data: map[string]interface{}{
 					"hello": "world",
 				},
@@ -213,7 +213,6 @@ func TestCodecV02_Decode(t *testing.T) {
 					Source:      *source,
 					ID:          "ABC-123",
 				},
-				DataEncoded: true,
 			},
 		},
 		"full v2 structured": {
