+++ pkg/cloudevents/transport/http/codec_v02_test.go
@@ -142,11 +142,11 @@ func TestCodecV02_Encode(t *testing.T) {
 		"simple v0.2 structured": {
 			codec: http.CodecV02{Encoding: http.StructuredV02},
 			event: cloudevents.Event{
-				Context: &cloudevents.EventContextV02{
+				Context: cloudevents.EventContextV02{
 					Type:   "com.example.test",
 					Source: *source,
 					ID:     "ABC-123",
-				},
+				}.AsV02(),
 			},
 			want: &http.Message{
 				Header: map[string][]string{
@@ -167,7 +167,7 @@ func TestCodecV02_Encode(t *testing.T) {
 		"full v0.2 structured": {
 			codec: http.CodecV02{Encoding: http.StructuredV02},
 			event: cloudevents.Event{
-				Context: &cloudevents.EventContextV02{
+				Context: cloudevents.EventContextV02{
 					ID:          "ABC-123",
 					Time:        &now,
 					Type:        "com.example.test",
@@ -177,7 +177,7 @@ func TestCodecV02_Encode(t *testing.T) {
 					Extensions: map[string]interface{}{
 						"test": "extended",
 					},
-				},
+				}.AsV02(),
 				Data: map[string]interface{}{
 					"hello": "world",
 				},
@@ -270,7 +270,6 @@ func TestCodecV02_Decode(t *testing.T) {
 					Source:      *source,
 					ID:          "ABC-123",
 				},
-				DataEncoded: true,
 			},
 		},
 		"full v0.2 binary": {
@@ -337,7 +336,6 @@ func TestCodecV02_Decode(t *testing.T) {
 					Source:      *source,
 					ID:          "ABC-123",
 				},
-				DataEncoded: true,
 			},
 		},
 		"full v0.2 structured": {
@@ -399,7 +397,6 @@ func TestCodecV02_Decode(t *testing.T) {
 					Source:      *source,
 					ID:          "ABC-123",
 				},
-				DataEncoded: true,
 			},
 		},
 	}
