+++ pkg/cloudevents/transport/nats/codec_v02_test.go
@@ -29,11 +29,11 @@ func TestCodecV02_Encode(t *testing.T) {
 		"simple v2 default": {
 			codec: nats.CodecV02{},
 			event: cloudevents.Event{
-				Context: &cloudevents.EventContextV02{
+				Context: cloudevents.EventContextV02{
 					Type:   "com.example.test",
 					Source: *source,
 					ID:     "ABC-123",
-				},
+				}.AsV02(),
 			},
 			want: &nats.Message{
 				Body: func() []byte {
@@ -51,7 +51,7 @@ func TestCodecV02_Encode(t *testing.T) {
 		"full v2 default": {
 			codec: nats.CodecV02{},
 			event: cloudevents.Event{
-				Context: &cloudevents.EventContextV02{
+				Context: cloudevents.EventContextV02{
 					ID:          "ABC-123",
 					Time:        &now,
 					Type:        "com.example.test",
@@ -61,7 +61,7 @@ func TestCodecV02_Encode(t *testing.T) {
 					Extensions: map[string]interface{}{
 						"test": "extended",
 					},
-				},
+				}.AsV02(),
 				Data: map[string]interface{}{
 					"hello": "world",
 				},
@@ -88,11 +88,11 @@ func TestCodecV02_Encode(t *testing.T) {
 		"simple v2 structured": {
 			codec: nats.CodecV02{Encoding: nats.StructuredV02},
 			event: cloudevents.Event{
-				Context: &cloudevents.EventContextV02{
+				Context: cloudevents.EventContextV02{
 					Type:   "com.example.test",
 					Source: *source,
 					ID:     "ABC-123",
-				},
+				}.AsV02(),
 			},
 			want: &nats.Message{
 				Body: func() []byte {
@@ -110,7 +110,7 @@ func TestCodecV02_Encode(t *testing.T) {
 		"full v2 structured": {
 			codec: nats.CodecV02{Encoding: nats.StructuredV02},
 			event: cloudevents.Event{
-				Context: &cloudevents.EventContextV02{
+				Context: cloudevents.EventContextV02{
 					ID:          "ABC-123",
 					Time:        &now,
 					Type:        "com.example.test",
@@ -120,7 +120,7 @@ func TestCodecV02_Encode(t *testing.T) {
 					Extensions: map[string]interface{}{
 						"test": "extended",
 					},
-				},
+				}.AsV02(),
 				Data: map[string]interface{}{
 					"hello": "world",
 				},
@@ -208,7 +208,6 @@ func TestCodecV02_Decode(t *testing.T) {
 					Source:      *source,
 					ID:          "ABC-123",
 				},
-				DataEncoded: true,
 			},
 		},
 		"full v2 structured": {
